* misc/strnlen.c: Cast pointer difference to size_t to avoid the compiler warning. * misc/mingw-aligned-malloc.c: Remove prototypes of __mingw_aligned_* now that they are in malloc.h and stdlib.h already includes malloc.h.
git-svn-id: svn+ssh://svn.code.sf.net/p/mingw-w64/code/branches/releases/v1.0@3584 4407c894-4637-0410-b4f5-ada5f102cad1

+++ mingw-w64-crt/misc/strnlen.c
@@ -4,7 +4,7 @@
 size_t __cdecl strnlen (const char *s, size_t maxlen)
 {
   const char *s2 = s;
-  while (s2 - s < maxlen && *s2)
+  while ((size_t) (s2 - s) < maxlen && *s2)
     ++s2;
   return s2 - s;
 }
